// src/enum.ts
// crud 增查改删 create  retrieve  update  delete

export const enum AUTHORITY {
	/**
	 *	增 人员
	 */
	PERSONNEL_C = 'PERSONNEL_C',
	/**
	 *	查 人员
	 */
	PERSONNEL_R = 'PERSONNEL_R',
	/**
	 *	改 人员
	 */
	PERSONNEL_U = 'PERSONNEL_U',
	/**
	 *	删 人员
	 */
	PERSONNEL_D = 'PERSONNEL_D',
	/**
	 * 产品
	 */
	PEUDUCT_C = 'PEUDUCT_C',
	/**产品*/
	PEUDUCT_R = 'PEUDUCT_R',
	/**产品*/
	PEUDUCT_U = 'PEUDUCT_U',
	/**产品*/
	PEUDUCT_D = 'PEUDUCT_D',
	/**角色*/
	ROLE_C = 'ROLE_C',
	/**角色*/
	ROLE_R = 'ROLE_R',
	/**角色*/
	ROLE_U = 'ROLE_U',
	/**角色*/
	ROLE_D = 'ROLE_D',

	/**地府*/
	HELL_R = 'HELL_R',
	/**1 */
	HELL_ONE_C = 'HELL_ONE_C',
	HELL_ONE_R = 'HELL_ONE_R',
	HELL_ONE_U = 'HELL_ONE_U',
	HELL_ONE_D = 'HELL_ONE_D',
	/**2 */
	HELL_TWO_C = 'HELL_TWO_C',
	HELL_TWO_R = 'HELL_TWO_R',
	HELL_TWO_U = 'HELL_TWO_U',
	HELL_TWO_D = 'HELL_TWO_D',
	/**3 */
	HELL_THREE_C = 'HELL_THREE_C',
	HELL_THREE_R = 'HELL_THREE_R',
	HELL_THREE_U = 'HELL_THREE_U',
	HELL_THREE_D = 'HELL_THREE_D',
	/**4 */
	HELL_FOUR_C = 'HELL_FOUR_C',
	HELL_FOUR_R = 'HELL_FOUR_R',
	HELL_FOUR_U = 'HELL_FOUR_U',
	HELL_FOUR_D = 'HELL_FOUR_D',
	/**5 */
	HELL_FIVES_C = 'HELL_FIVES_C',
	HELL_FIVES_R = 'HELL_FIVES_R',
	HELL_FIVES_U = 'HELL_FIVES_U',
	HELL_FIVES_D = 'HELL_FIVES_D',
	/**6 */
	HELL_SIXTY_C = 'HELL_SIXTY_C',
	HELL_SIXTY_R = 'HELL_SIXTY_R',
	HELL_SIXTY_U = 'HELL_SIXTY_U',
	HELL_SIXTY_D = 'HELL_SIXTY_D',
	/**7 */
	HELL_SEVEN_C = 'HELL_SEVEN_C',
	HELL_SEVEN_R = 'HELL_SEVEN_R',
	HELL_SEVEN_U = 'HELL_SEVEN_U',
	HELL_SEVEN_D = 'HELL_SEVEN_D',
	/**8 */
	HELL_EIGHT_C = 'HELL_EIGHT_C',
	HELL_EIGHT_R = 'HELL_EIGHT_R',
	HELL_EIGHT_U = 'HELL_EIGHT_U',
	HELL_EIGHT_D = 'HELL_EIGHT_D',
	/**9 */
	HELL_NINE_C = 'HELL_NINE_C',
	HELL_NINE_R = 'HELL_NINE_R',
	HELL_NINE_U = 'HELL_NINE_U',
	HELL_NINE_D = 'HELL_NINE_D',
	/**10 */
	HELL_TEN_C = 'HELL_TEN_C',
	HELL_TEN_R = 'HELL_TEN_R',
	HELL_TEN_U = 'HELL_TEN_U',
	HELL_TEN_D = 'HELL_TEN_D',
	/**11 */
	HELL_ELEVEN_C = 'HELL_ELEVEN_C',
	HELL_ELEVEN_R = 'HELL_ELEVEN_R',
	HELL_ELEVEN_U = 'HELL_ELEVEN_U',
	HELL_ELEVEN_D = 'HELL_ELEVEN_D',
	/**12 */
	HELL_TWELVE_C = 'HELL_TWELVE_C',
	HELL_TWELVE_R = 'HELL_TWELVE_R',
	HELL_TWELVE_U = 'HELL_TWELVE_U',
	HELL_TWELVE_D = 'HELL_TWELVE_D',
	/**13 */
	HELL_THIRTEEN_C = 'HELL_THIRTEEN_C',
	HELL_THIRTEEN_R = 'HELL_THIRTEEN_R',
	HELL_THIRTEEN_U = 'HELL_THIRTEEN_U',
	HELL_THIRTEEN_D = 'HELL_THIRTEEN_D',
	/**14 */
	HELL_FOURTEEN_C = 'HELL_FOURTEEN_C',
	HELL_FOURTEEN_R = 'HELL_FOURTEEN_R',
	HELL_FOURTEEN_U = 'HELL_FOURTEEN_U',
	HELL_FOURTEEN_D = 'HELL_FOURTEEN_D',
	/**15 */
	HELL_FIFTEEN_C = 'HELL_FIFTEEN_C',
	HELL_FIFTEEN_R = 'HELL_FIFTEEN_R',
	HELL_FIFTEEN_U = 'HELL_FIFTEEN_U',
	HELL_FIFTEEN_D = 'HELL_FIFTEEN_D',
	/**16 */
	HELL_SIXTEEN_C = 'HELL_SIXTEEN_C',
	HELL_SIXTEEN_R = 'HELL_SIXTEEN_R',
	HELL_SIXTEEN_U = 'HELL_SIXTEEN_U',
	HELL_SIXTEEN_D = 'HELL_SIXTEEN_D',
	/**17 */
	HELL_SEVENTEEN_C = 'HELL_SEVENTEEN_C',
	HELL_SEVENTEEN_R = 'HELL_SEVENTEEN_R',
	HELL_SEVENTEEN_U = 'HELL_SEVENTEEN_U',
	HELL_SEVENTEEN_D = 'HELL_SEVENTEEN_D',
	/**18 */
	HELL_EIGHTEEN_C = 'HELL_EIGHTEEN_C',
	HELL_EIGHTEEN_R = 'HELL_EIGHTEEN_R',
	HELL_EIGHTEEN_U = 'HELL_EIGHTEEN_U',
	HELL_EIGHTEEN_D = 'HELL_EIGHTEEN_D',
}

export const enum ROLE {
	ROOTADMIN = 'ROOTADMIN',
}
/*





*/